Added the 'play page' and some functions & btns
In this commit I added the 'play page' which inclues a startSequence function which loops over and array including all abc words, and return a random letter, also all of this is wrapped in an interval so the user can visualize the loop. 

In addition, when the loop finishes, I added a restart btn that has an event listener who trigger the 'startFunction' again.